feat: adicionar atualizações no backend + atualização do README

+++ projetoweb2/src/main/java/com/bti/projetoweb2/visitas/repositories/VisitaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.bti.projetoweb2.visitas.entities.Visita;
 
-public interface VisitaRepository extends JpaRepository<Visita, Long> {
+public interface VisitaRepository extends JpaRepository<Visita, Integer> {
 
     // Busca por nome (ignorando maiúsculas/minúsculas)
     List<Visita> findByNomeContainingIgnoreCase(String parteNome);
